avionics and garmin files separated
