--- frontend/src/components/Avionics/AvionicsAd.css
@@ -1,3 +1,0 @@
-.AvionicsAd_jumbo{
-    height: 150px;
-}
