+++ frontend/src/components/Avionics/AvionicsImg.js
@@ -1,6 +1,6 @@
 import {Jumbotron} from "reactstrap";
 import {Button, Card, Col, Row} from "react-bootstrap";
-import AvionicsProducts from "./AvionicsProducts";
+import GarminBanner from "../GarminBanner/GarminBanner";
 import React from "react";
 import Zoom from 'react-reveal/Zoom'
 
@@ -188,7 +188,7 @@ const AvionicsImg = () => {
                         </Row>
                     </Col>
 
-                    <AvionicsProducts/>
+                    <GarminBanner/>
 
                 </Row>
             </Jumbotron>
