+++ frontend/src/components/GarminBanner/GarminHomeFlag.scss
