+++ frontend/src/components/GarminBanner/ProductsCarouselGarminAd.js
@@ -1,16 +1,16 @@
 import React from "react";
 import Fade from 'react-reveal/Fade';
 import {Col, Row} from "react-bootstrap";
-import AvionicsProducts from "../Avionics/AvionicsProducts";
+import GarminBanner from "./GarminBanner";
 import './GarminHomeFlag.scss'
-import ProductCarousel from "./ProductCarousel";
+import ProductCarousel from "../ProductsCarousel/ProductCarousel";
 
 
 const ProductsCarouselGarminAd = () => {
     return (
         <Row className='d-flex justify-content-center Garmin_card-row bg-white'>
             <Col md={6} xs={12} className='d-flex align-content-center m-0 p-0'>
-                <AvionicsProducts/>
+                <GarminBanner/>
             </Col>
             <Col md={6} xs={12} className='bg-white d-flex justify-content-center m-0 p-0'>
                 <Fade duration={1500} right>
